Clean up code for navigation bar
