+++ src/components/marginals/index.tsx
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export {default as Header} from './Header';
+export {default as Navbar} from './Navbar';
