--- src/utils/scrollReveal.ts
@@ -1,6 +1,0 @@
-// import ScrollReveal from 'scrollreveal';
-
-// const isSSR = typeof window === 'undefined';
-// const sr = isSSR ? null : ScrollReveal();
-
-// export default sr;
